pyproject.toml
# PyProject configuration for Satellite Thruster Control System

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "satellite-thruster-control"
description = "Model Predictive Control system for satellite thruster control"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Aevar Oefjoerd"}
]
maintainers = [
    {name = "Aevar Oefjoerd"}
]
keywords = ["satellite", "control", "mpc", "aerospace", "simulation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "PyQt5>=5.15.0",
    "opencv-python>=4.5.0",
    "gurobipy>=11.0.0",
    "pyserial>=3.5",
    "requests>=2.26.0",
    "ezdxf>=0.17.0",
    "shapely>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "pytest-timeout>=2.1",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "bandit>=1.7",
    "safety>=2.3",
]

[project.urls]
Homepage = "https://github.com/AevarOfjord/SatelliteProject"
Documentation = "https://github.com/AevarOfjord/SatelliteProject"
Repository = "https://github.com/AevarOfjord/SatelliteProject"
"Bug Tracker" = "https://github.com/AevarOfjord/SatelliteProject/issues"

[project.scripts]
satellite-sim = "simulation:main"
satellite-real = "real:main"
satellite-visualize = "visualization_manager:main"
satellite-test = "testing_environment:main"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | Data
  | DXF
  | Motive
  | Thruster_Test
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["Data/*", "DXF/*", "Motive/*", "Thruster_Test/*", "venv/*"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (moderate speed)",
    "hardware: Hardware-dependent tests (requires physical setup)",
    "slow: Slow tests (>1 second)",
]

# Coverage configuration
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "setup.py",
    "*/Data/*",
    "*/DXF/*",
    "*/Motive/*",
    "*/Thruster_Test/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["tests", "Data", "DXF", "Motive", "Thruster_Test"]
skips = ["B101", "B601"]  # Skip assert warnings and shell injection (if needed)

# Mypy type checking configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["Data", "DXF", "Motive", "Thruster_Test", "venv", "build"]
